refactor(query): remove legacy aggregate hash index
