+++ src/query/expression/src/aggregate/aggregate_meta.rs
@@ -45,15 +45,12 @@ impl SerializedPayload {
         aggrs: Vec<Arc<dyn AggregateFunction>>,
         num_states: usize,
         radix_bits: u64,
-        enable_experiment_hash_index: bool,
         arena: Arc<Bump>,
         need_init_entry: bool,
     ) -> Result<AggregateHashTable> {
         let rows_num = self.data_block.num_rows();
         let capacity = AggregateHashTable::get_capacity_for_count(rows_num);
-        let config = HashTableConfig::default()
-            .with_initial_radix_bits(radix_bits)
-            .with_experiment_hash_index(enable_experiment_hash_index);
+        let config = HashTableConfig::default().with_initial_radix_bits(radix_bits);
         let mut state = ProbeState::default();
         let group_len = group_types.len();
         let mut hashtable = AggregateHashTable::new_directly(
@@ -89,7 +86,6 @@ impl SerializedPayload {
         aggrs: Vec<Arc<dyn AggregateFunction>>,
         num_states: usize,
         radix_bits: u64,
-        enable_experiment_hash_index: bool,
         arena: Arc<Bump>,
     ) -> Result<PartitionedPayload> {
         let hashtable = self.convert_to_aggregate_table(
@@ -97,7 +93,6 @@ impl SerializedPayload {
             aggrs,
             num_states,
             radix_bits,
-            enable_experiment_hash_index,
             arena,
             false,
         )?;
@@ -109,18 +104,10 @@ impl SerializedPayload {
         group_types: Vec<DataType>,
         aggrs: Vec<Arc<dyn AggregateFunction>>,
         num_states: usize,
-        enable_experiment_hash_index: bool,
         arena: Arc<Bump>,
     ) -> Result<Payload> {
-        let hashtable = self.convert_to_aggregate_table(
-            group_types,
-            aggrs,
-            num_states,
-            0,
-            enable_experiment_hash_index,
-            arena,
-            false,
-        )?;
+        let hashtable =
+            self.convert_to_aggregate_table(group_types, aggrs, num_states, 0, arena, false)?;
         Ok(hashtable.payload.into_single_payload())
     }
 
@@ -130,7 +117,6 @@ impl SerializedPayload {
         group_types: Vec<DataType>,
         aggrs: Vec<Arc<dyn AggregateFunction>>,
         num_states: usize,
-        enable_experiment_hash_index: bool,
         flush_state: &mut PayloadFlushState,
     ) -> Result<Vec<AggregatePayload>> {
         Ok(self
@@ -139,7 +125,6 @@ impl SerializedPayload {
                 aggrs,
                 num_states,
                 0,
-                enable_experiment_hash_index,
                 Arc::new(Bump::new()),
                 false,
             )?
