+++ src/query/expression/src/aggregate/hash_index/index.rs
@@ -16,14 +16,14 @@
 use std::hint::likely;
 use std::mem::size_of;
 
-use crate::aggregate::NEW_INDEX_LOAD_FACTOR;
+use crate::aggregate::HASH_INDEX_LOAD_FACTOR;
 use crate::aggregate::ProbeState;
-use crate::aggregate::legacy_hash_index::TableAdapter;
-use crate::aggregate::new_hash_index::bitmask::Tag;
-use crate::aggregate::new_hash_index::group::Group;
+use crate::aggregate::hash_index::bitmask::Tag;
+use crate::aggregate::hash_index::group::Group;
+use crate::aggregate::hash_index_adapter::TableAdapter;
 use crate::aggregate::row_ptr::RowPtr;
 
-pub struct ExperimentalHashIndex {
+pub(in crate::aggregate) struct HashIndex {
     ctrls: Vec<Tag>,
     pointers: Vec<RowPtr>,
     capacity: usize,
@@ -31,7 +31,7 @@ pub struct ExperimentalHashIndex {
     count: usize,
 }
 
-impl ExperimentalHashIndex {
+impl HashIndex {
     pub fn with_capacity(capacity: usize) -> Self {
         debug_assert!(capacity.is_power_of_two());
         // avoid handling: SMALL TABLE NASTY CORNER CASE
@@ -60,7 +60,7 @@ impl ExperimentalHashIndex {
     }
 }
 
-impl ExperimentalHashIndex {
+impl HashIndex {
     #[inline]
     fn set_ctrl(&mut self, index: usize, tag: Tag) {
         // Mirror: keep the tail padding in sync with the head so that
@@ -215,7 +215,7 @@ impl ExperimentalHashIndex {
     }
 }
 
-impl ExperimentalHashIndex {
+impl HashIndex {
     pub fn capacity(&self) -> usize {
         self.capacity
     }
@@ -225,7 +225,7 @@ impl ExperimentalHashIndex {
     }
 
     pub fn resize_threshold(&self) -> usize {
-        (self.capacity as f64 / NEW_INDEX_LOAD_FACTOR) as usize
+        (self.capacity as f64 / HASH_INDEX_LOAD_FACTOR) as usize
     }
 
     pub fn allocated_bytes(&self) -> usize {
@@ -245,5 +245,160 @@ impl ExperimentalHashIndex {
         let index = self.probe_empty(hash);
         self.pointers[index] = row_ptr;
         self.count += 1;
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::collections::HashMap;
+
+    use super::*;
+
+    struct TestTableAdapter {
+        incoming: Vec<(u64, u64)>,     // (key, hash)
+        payload: Vec<(u64, u64, u64)>, // (key, hash, value)
+        init_count: usize,
+        pin_data: Box<[u8]>,
+    }
+
+    impl TestTableAdapter {
+        fn new(incoming: Vec<(u64, u64)>, payload: Vec<(u64, u64, u64)>) -> Self {
+            Self {
+                incoming,
+                init_count: payload.len(),
+                payload,
+                pin_data: vec![0; 1000].into(),
+            }
+        }
+
+        fn init_state(&self) -> ProbeState {
+            let mut state = ProbeState {
+                row_count: self.incoming.len(),
+                ..Default::default()
+            };
+
+            for (i, (_, hash)) in self.incoming.iter().enumerate() {
+                state.group_hashes[i] = *hash
+            }
+
+            state
+        }
+
+        fn init_hash_index(&self, hash_index: &mut HashIndex) {
+            for (i, (_, hash, _)) in self.payload.iter().copied().enumerate() {
+                hash_index.probe_slot_and_set(hash, self.get_row_ptr(false, i));
+            }
+        }
+
+        fn get_row_ptr(&self, incoming: bool, row: usize) -> RowPtr {
+            RowPtr::new(unsafe {
+                self.pin_data
+                    .as_ptr()
+                    .add(if incoming { row + self.init_count } else { row })
+            })
+        }
+
+        fn get_payload(&self, row_ptr: RowPtr) -> (u64, u64, u64) {
+            let index = row_ptr.as_ptr() as usize - self.pin_data.as_ptr() as usize;
+            self.payload[index]
+        }
+    }
+
+    impl TableAdapter for TestTableAdapter {
+        fn append_rows(&mut self, state: &mut ProbeState, new_entry_count: usize) {
+            for row in state.empty_vector[..new_entry_count].iter() {
+                let (key, hash) = self.incoming[*row];
+                let value = key + 20;
+
+                self.payload.push((key, hash, value));
+                state.addresses[*row] = self.get_row_ptr(true, row.to_usize());
+            }
+        }
+
+        fn compare(
+            &mut self,
+            state: &mut ProbeState,
+            need_compare_count: usize,
+            mut no_match_count: usize,
+        ) -> usize {
+            for row in state.group_compare_vector[..need_compare_count]
+                .iter()
+                .copied()
+            {
+                let incoming = self.incoming[row];
+                let (key, _, _) = self.get_payload(state.addresses[row]);
+
+                if incoming.0 == key {
+                    continue;
+                }
+
+                state.no_match_vector[no_match_count] = row;
+                no_match_count += 1;
+            }
+
+            no_match_count
+        }
+    }
+
+    #[derive(Clone)]
+    struct TestCase {
+        capacity: usize,
+        incoming: Vec<(u64, u64)>,     // (key, hash)
+        payload: Vec<(u64, u64, u64)>, // (key, hash, value)
+        want_count: usize,
+        want: HashMap<u64, u64>,
+    }
+
+    impl TestCase {
+        fn run_hash_index(self) {
+            let TestCase {
+                capacity,
+                incoming,
+                payload,
+                want_count,
+                want,
+            } = self;
+            let mut hash_index = HashIndex::with_capacity(capacity);
+            let mut adapter = TestTableAdapter::new(incoming, payload);
+            let mut state = adapter.init_state();
+
+            adapter.init_hash_index(&mut hash_index);
+
+            let count =
+                hash_index.probe_and_create(&mut state, adapter.incoming.len(), &mut adapter);
+
+            assert_eq!(want_count, count);
+
+            let got = state.addresses[..state.row_count]
+                .iter()
+                .map(|row_ptr| {
+                    let (key, _, value) = adapter.get_payload(*row_ptr);
+                    (key, value)
+                })
+                .collect::<HashMap<_, _>>();
+
+            assert_eq!(want, got);
+        }
+    }
+
+    #[test]
+    fn test_hash_index() {
+        TestCase {
+            capacity: 16,
+            incoming: vec![(1, 123), (2, 456), (3, 123), (4, 44)],
+            payload: vec![(4, 44, 77)],
+            want_count: 3,
+            want: HashMap::from_iter([(1, 21), (2, 22), (3, 23), (4, 77)]),
+        }
+        .run_hash_index();
+
+        TestCase {
+            capacity: 16,
+            incoming: vec![(1, 11 << 48), (2, 22 << 48), (3, 33 << 48), (4, 44 << 48)],
+            payload: vec![(4, 44 << 48, 77)],
+            want_count: 3,
+            want: HashMap::from_iter([(1, 21), (2, 22), (3, 23), (4, 77)]),
+        }
+        .run_hash_index();
     }
 }
