+++ src/query/service/src/pipelines/processors/transforms/aggregator/new_aggregate/new_transform_final_aggregate.rs
@@ -165,8 +165,7 @@ impl NewTransformFinalAggregate {
                 .with_partition_start_bit(Self::partition_start_bit(
                     base_consumed_bits,
                     spilled_depth,
-                ))
-                .with_experiment_hash_index(params.enable_experiment_hash_index),
+                )),
             Arc::new(Bump::new()),
         )
     }
@@ -211,7 +210,6 @@ impl NewTransformFinalAggregate {
             self.params.aggregate_functions.clone(),
             self.params.num_states(),
             0,
-            self.params.enable_experiment_hash_index,
             Arc::new(Bump::new()),
         )?;
