+++ src/query/service/src/pipelines/processors/transforms/aggregator/transform_aggregate_final.rs
@@ -129,7 +129,6 @@ impl TransformFinalAggregate {
                     self.params.aggregate_functions.clone(),
                     self.params.num_states(),
                     0,
-                    self.params.enable_experiment_hash_index,
                     Arc::new(Bump::new()),
                 )?;
                 ht.combine_payloads(&payload, &mut self.flush_state)?;
@@ -141,7 +140,6 @@ impl TransformFinalAggregate {
                     self.params.aggregate_functions.clone(),
                     self.params.num_states(),
                     0,
-                    self.params.enable_experiment_hash_index,
                     Arc::new(Bump::new()),
                     true,
                 )?);
@@ -167,9 +165,7 @@ impl TransformFinalAggregate {
                 let mut hashtable = AggregateHashTable::new_with_capacity(
                     self.params.group_data_types.clone(),
                     self.params.aggregate_functions.clone(),
-                    HashTableConfig::default()
-                        .with_initial_radix_bits(0)
-                        .with_experiment_hash_index(self.params.enable_experiment_hash_index),
+                    HashTableConfig::default().with_initial_radix_bits(0),
                     capacity,
                     Arc::new(Bump::new()),
                 );
