+++ src/query/settings/src/settings_default.rs
@@ -1721,13 +1721,6 @@ impl DefaultSettings {
                     scope: SettingScope::Both,
                     range: Some(SettingRange::String(vec!["auto".into(),"row".into(), "bucket".into()])),
                 }),
-                ("enable_experiment_hash_index", DefaultSettingValue {
-                    value: UserSettingValue::UInt64(1),
-                    desc: "experiment setting enable hash index(enabled by default).",
-                    mode: SettingMode::Both,
-                    scope: SettingScope::Both,
-                    range: Some(SettingRange::Numeric(0..=1)),
-                }),
             ]);
 
             Ok(Arc::new(DefaultSettings {
